requirements.txt
fastapi>=0.110
uvicorn[standard]
pydantic<3
python-dotenv
requests
